+++ core/core/sim/memory/ram/sparse.hpp
@@ -63,8 +63,8 @@ public:
 
 private:
   mutable struct PerformanceCounters {
-    u32 rd_bytes = 0;
-    u32 wr_bytes = 0;
+    u64 rd_bytes = 0;
+    u64 wr_bytes = 0;
   } _counters = {};
   Configuration _config;
   pepp::bts::PagedPool<u8> _pool;
